rearrange functions to avoid circular imports and copying code back and forth. scene interpolation (without time interpolation) *should* work time interpolation *does* work shared memory scene interpolation *doesn't* work.
